Codex - Issue #5 - Refine auth boundaries and session management
Decouple the auth service from MVC request DTOs with application-level command models, add a DI-backed current-user accessor, and extract client auth transport/session logic out of App.tsx into dedicated API and hook modules.
