+++ ProjectPortfolio2026/ProjectPortfolio2026.Server/Services/ServiceModels/AccountPasswordChangeCommand.cs
@@ -1,0 +1,8 @@
+namespace ProjectPortfolio2026.Server.Services.ServiceModels;
+
+public sealed class AccountPasswordChangeCommand
+{
+    public string CurrentPassword { get; init; } = string.Empty;
+
+    public string NewPassword { get; init; } = string.Empty;
+}
